Add power series
I've added base code for computing a power series expansion for a modular form on X_0(43) at the rational CM point \tau = (43 + sqrt(-43))/2. 
I have a type called "XuSeries" that tries to dynamically update the coefficients computed.
There may very well be errors in the code! This is just a first commit!